Started a section outlining the proof of the coherence property of monoidal categories.
I start by defining monoidal and symmetric monoidal categories.  Then
I define the free monoidal category in the style of Mac Lane using the
language of binary words, and show that there is a bijection between
the category of binary words and an arbitrary monoidal category.
Thus, there is a unique functor between these two categories. This
result implies that W is the free monoidal category.

git-svn-id: https://svn.divms.uiowa.edu/repos/clc/reports/harley_thesis@13678 bd178f5f-7bea-4014-844f-495ca19b8805