Just about down with an example program in Freedom of Speech that suppose we have the div and isZero functions, and then defines a new function called div-plus-2 which takes in two natural numbers and adds two to the first number and then divides that by the second argument, however, I use abort to abort the computation when the second argument is zero.
To abort the computation I wrap the term abort in a lambda-abstraction,
and then I defined an abstraction of the Church-encoded booleans. The
abstraction is instead of just returning one of the two arguments in the
definition of true and false I allow a function to be applied to the
arguments.  So if isZero returns true -- of the abstracted booleans --
then I call true with my abort lambda.  Thus, the computation will
abort.  If it returns false, then I pass the identity function to false
along with the result of the division.

If anyone is interested see Example 7.0.1.18 on page 129 of the PDF.

So this is more progress on #1.